Split merchant services
